requirements.txt
aiogram==3.3.0
apscheduler==3.10.4
Django==4.2.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.13.1 